Move event map to separate file
